requirements.txt
openai>=1.12.0
pyodbc>=5.0.1
python-dotenv>=1.0.0